Refactor pixel size observer
